backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
bcrypt==4.1.2
aiohttp==3.9.1
starlette==0.37.2
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.12.0